fix: UserContext.jsx, Search 컴포넌트, Header 컴포넌트 타입에러 수정
