+++ src/components/Search/Search.styled.ts
